+++ src/Common/ThreadStatus.cpp
@@ -4,6 +4,7 @@
 #include <Common/ThreadStatus.h>
 #include <base/errnoToString.h>
 #include <Interpreters/OpenTelemetrySpanLog.h>
+#include <Interpreters/Context.h>
 
 #include <Poco/Logger.h>
 #include <base/getThreadId.h>
@@ -11,6 +12,7 @@
 
 #include <csignal>
 #include <mutex>
+#include <sys/mman.h>
 
 
 namespace DB
